streaming/requirements.txt
pymongo
kafka-python
dnspython